WV-2790 sidebar buttons (#4554)
* Full Screen Styling back towards normal

* Mobile Styling back towards normal

* Chart mode styling done. Chart mode is no longer available in mobile mode.

* Lint fixes

* Merge conflict cleanup
